--- app/src/main/java/com/kuit/ourmenu/ui/home/component/HomeSubRecommendationText.kt
@@ -1,20 +1,0 @@
-package com.kuit.ourmenu.ui.home.component
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.kuit.ourmenu.R
-import com.kuit.ourmenu.ui.theme.Neutral900
-import com.kuit.ourmenu.ui.theme.ourMenuTypography
-
-
